Variable delta en PumpController

+++ main/WaterTank.cpp
@@ -59,7 +59,7 @@ void WaterTank::setMinWaterDistance(unsigned int min_water_distance) {
     this->min_water_distance = min_water_distance;
 }
 
-uint8_t WaterTank::getWaterTankLevel() {
+int WaterTank::getWaterTankLevel() {
     return water_level;
 }
